+++ src/index.ts
@@ -1,5 +1,6 @@
 import { AppWrapperRoute, defineWebApplication } from '@ownclouders/web-pkg'
 import App from './App.vue'
+import { mimeTypes } from './mimeTypes'
 import { id as appId } from '../public/manifest.json'
 
 export default defineWebApplication({
@@ -26,27 +27,12 @@ export default defineWebApplication({
       appInfo: {
         name: '3D Model Viewer',
         id: appId,
-        icon: 'resource-type-graphic',
-        iconFillType: 'fill',
-        iconColor: '#86C540',
-        extensions: [
-          {
-            extension: 'glb',
-            label: 'View 3D Model'
-          },
-          {
-            extension: 'stl',
-            label: 'View 3D Model'
-          },
-          {
-            extension: 'fbx',
-            label: 'View 3D Model'
-          },
-          {
-            extension: 'obj',
-            label: 'View 3D Model'
-          }
-        ]
+        img: new URL('./assets/images/3d-model-viewer.svg', import.meta.url).href,
+        extensions: Object.keys(mimeTypes).map((extension) => ({
+          extension,
+          mimeType: mimeTypes[extension],
+          label: 'View in 3D Model Viewer'
+        }))
       },
       routes
     }