+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+Added App Icon draft
 Added basic OBJ support
 Added light and ambient
 Added default material
@@ -15,8 +16,14 @@ Added basic FBX support
 Added basic debugging mode
 Added basic STL support 
 
+- Support `.obj` file format
+- Support `.fbx` file format
+- Support `.stl` file format
+
 ### Fixed
 Fixed unloading models on prev/next
+
+- Unloading models on prev/next
 
 ### Changed